requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.6
PyJWT>=2.8.0
bcrypt>=4.1.2
pycryptodome>=3.20.0
python-dotenv>=1.0.0
# For production WSGI:
# gunicorn>=21.2.0
# gevent>=24.2.1
# gevent-websocket>=0.10.1
